requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
websockets>=13.0
httpx>=0.27.0
openai>=1.50.0
anthropic>=0.39.0
python-dotenv>=1.0.0
pydantic>=2.9.0
mcp>=1.27.0
